- Renamed the folder "infrastructure" to "api_infrastructure"
